Overlay - convert to Strict ESM (#344)
