+++ overlay/components/RuntimeErrorStack.js
@@ -1,6 +1,6 @@
-const ErrorStackParser = require('error-stack-parser');
-const theme = require('../theme');
-const formatFilename = require('../utils/formatFilename');
+import * as ErrorStackParser from 'error-stack-parser';
+import theme from '../theme.js';
+import { formatFilename } from '../utils.js';
 
 /**
  * @typedef {Object} RuntimeErrorStackProps
@@ -76,4 +76,4 @@ function RuntimeErrorStack(document, root, props) {
   root.appendChild(stackContainer);
 }
 
-module.exports = RuntimeErrorStack;
+export default RuntimeErrorStack;
